@@ -1,11 +1,12 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
-    <TargetFramework>netcoreapp2.1</TargetFramework>
+	  <TargetFramework>net6.0</TargetFramework>
+	  <Version>6.0.0</Version>
   </PropertyGroup>
 
   <PropertyGroup Condition="'$(Configuration)|$(Platform)'=='Debug|AnyCPU'">
-    <DocumentationFile>bin\Debug\netcoreapp2.1\Ae.AccountAuthority.Service.Core.xml</DocumentationFile>
+    <DocumentationFile>bin\Debug\net6.0\Ae.AccountAuthority.Service.Core.xml</DocumentationFile>
   </PropertyGroup>
 
   <ItemGroup>
