@@ -26,6 +26,9 @@ using Ae.Vehicle.Service.Imp.Services;
 using Ae.Vehicle.Service.Validators;
 using ApolloErp.Redis;
 using Ae.Vehicle.Service.Filters;
+using Ae.Vehicle.Service.Common.Format;
+using Microsoft.Extensions.Hosting;
+using Microsoft.AspNetCore.Http;
 
 namespace Ae.Vehicle.Service
 {
@@ -41,16 +44,30 @@ namespace Ae.Vehicle.Service
 
         public void ConfigureServices(IServiceCollection services)
         {
-            services.AddMvc()
-                .AddJsonOptions(options => { options.SerializerSettings.DateFormatString = "yyyy-MM-dd HH:mm:ss"; })
-                .SetCompatibilityVersion(CompatibilityVersion.Version_2_1).AddFluentValidation(o =>
+            //services.AddMvc()
+            //    .AddJsonOptions(options => { options.SerializerSettings.DateFormatString = "yyyy-MM-dd HH:mm:ss"; })
+            //    .SetCompatibilityVersion(CompatibilityVersion.Version_2_1).AddFluentValidation(o =>
+            //    {
+            //        o.RegisterValidatorsFromAssemblyContaining<GetVehicleListByBrandRequestValidator>();
+            //        o.RegisterValidatorsFromAssemblyContaining<PaiLiangByVehicleIdRequestValidator>();
+            //        o.RegisterValidatorsFromAssemblyContaining<VehicleNianByPaiLiangRequestValidator>();
+            //        o.RegisterValidatorsFromAssemblyContaining<VehicleSalesNameByNianRequestValidator>();
+            //    });
+            //services.AddControllers()
+            //.AddJsonOptions(options => { options.JsonSerializerOptions.Converters.Add(new DatetimeJsonConverter("yyyy-MM-dd HH:mm:ss")); });
+            services.AddMvc(options => {
+                options.EnableEndpointRouting = false;  //关闭Endpoint的路由支持来兼容
+                options.SuppressAsyncSuffixInActionNames = false;  //关闭新特性：Async结尾会默认去除
+            })
+                .AddNewtonsoftJson()
+                .AddJsonOptions(options => { options.JsonSerializerOptions.Converters.Add(new DatetimeJsonConverter("yyyy-MM-dd HH:mm:ss")); })
+                .AddFluentValidation(o =>
                 {
                     o.RegisterValidatorsFromAssemblyContaining<GetVehicleListByBrandRequestValidator>();
                     o.RegisterValidatorsFromAssemblyContaining<PaiLiangByVehicleIdRequestValidator>();
                     o.RegisterValidatorsFromAssemblyContaining<VehicleNianByPaiLiangRequestValidator>();
                     o.RegisterValidatorsFromAssemblyContaining<VehicleSalesNameByNianRequestValidator>();
                 });
-
             // override modelstate
             services.Configure<ApiBehaviorOptions>(options =>
             {
@@ -120,9 +137,10 @@ namespace Ae.Vehicle.Service
             });
         }
 
-        public void Configure(IApplicationBuilder app, IHostingEnvironment env)
+        public void Configure(IApplicationBuilder app, IWebHostEnvironment env)
         {
-            if (!env.IsProduction())
+            //app.UseRouting();
+            if (env.IsDevelopment())
             {
                 app.UseDeveloperExceptionPage();
 
@@ -156,6 +174,17 @@ namespace Ae.Vehicle.Service
                     "default",
                     "{controller=Home}/{action=Index}");
             });
+            //app.UseEndpoints(endpoints =>
+            //{
+            //    // 设置默认路由
+            //    endpoints.MapGet("/", async context =>
+            //    {
+            //        await context.Response.WriteAsync("Hello, World!");
+            //    });
+
+            //    // 配置控制器路由
+            //    endpoints.MapControllers();
+            //});
         }
     }
 }
