@@ -92,7 +92,7 @@ export {
 
   export function LoginAsync(data) {
     return request2({
-      url: `${domain2}Authentication/LoginAsync`,
+      url: `${domain2}/Authentication/LoginAsync`,
       method: 'post',
       data
     })
@@ -117,7 +117,7 @@ export {
 
   export function GetOrgPageListByAuthCodeAsync(data) {
     return request2({
-      url: `${domain2}Authentication/GetOrgPageListByAuthCodeAsync`,
+      url: `${domain2}/Authentication/GetOrgPageListByAuthCodeAsync`,
       method: 'post',
       data
     })
@@ -125,7 +125,7 @@ export {
 
   export function GetTokenInfoByAuthCode(data) {
     return request2({
-      url: `${domain2}Authentication/GetTokenInfoByAuthCode`,
+      url: `${domain2}/Authentication/GetTokenInfoByAuthCode`,
       method: 'post',
       data
     })
@@ -133,7 +133,7 @@ export {
 
   export function GetSMSCodeAsync(data) {
     return request2({
-      url: `${domain2}Authentication/GetSMSCodeAsync`,
+      url: `${domain2}/Authentication/GetSMSCodeAsync`,
       method: 'post',
       data
     })
@@ -141,7 +141,7 @@ export {
 
   export function LoginWithSMSAsync(data) {
     return request2({
-      url: `${domain2}Authentication/LoginWithSMSAsync`,
+      url: `${domain2}/Authentication/LoginWithSMSAsync`,
       method: 'post',
       data
     })
