@@ -103,7 +103,7 @@ export {
 
 export function LoginAsync(data) {
   return request2({
-    url: `${domain2}Authentication/LoginAsync`,
+    url: `${domain2}/Authentication/LoginAsync`,
     method: 'post',
     data
   })
@@ -128,7 +128,7 @@ export function logout() {
 
 export function GetOrgPageListByAuthCodeAsync(data) {
   return request2({
-    url: `${domain2}Authentication/GetOrgPageListByAuthCodeAsync`,
+    url: `${domain2}/Authentication/GetOrgPageListByAuthCodeAsync`,
     method: 'post',
     data
   })
@@ -136,7 +136,7 @@ export function GetOrgPageListByAuthCodeAsync(data) {
 
 export function GetTokenInfoByAuthCode(data) {
   return request2({
-    url: `${domain2}Authentication/GetTokenInfoByAuthCode`,
+    url: `${domain2}/Authentication/GetTokenInfoByAuthCode`,
     method: 'post',
     data
   })
@@ -144,7 +144,7 @@ export function GetTokenInfoByAuthCode(data) {
 
 export function GetSMSCodeAsync(data) {
   return request2({
-    url: `${domain2}Authentication/GetSMSCodeAsync`,
+    url: `${domain2}/Authentication/GetSMSCodeAsync`,
     method: 'post',
     data
   })
@@ -152,14 +152,14 @@ export function GetSMSCodeAsync(data) {
 
 export function LoginWithSMS(data) {
   return request2({
-    url: `${domain2}Authentication/LoginWithSMSAsync`,
+    url: `${domain2}/Authentication/LoginWithSMSAsync`,
     method: 'post',
     data
   })
 }
 export function GetEmpAuthorityListForWebByEmpIdAsync(params) {
   return request2({
-    url: `${domain1}Authorization/GetEmpAuthorityListForWebByEmpIdAsync`,
+    url: `${domain1}/Authorization/GetEmpAuthorityListForWebByEmpIdAsync`,
     method: 'get',
     params
   })
