@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using System.Text;
 using Newtonsoft.Json.Converters;
 
-namespace Ae.Shop.Api.Common.Format.DateTime
+namespace Ae.Shop.Api.Common.Format
 {
     public class FmtDTToYearMonthDayLowerWithStrikethrough : IsoDateTimeConverter
     {
