@@ -214,7 +214,7 @@ export {
 
   export function LoginAsync(data) {
     return request2({
-      url: `${domain2}Authentication/LoginAsync`,
+      url: `${domain2}/Authentication/LoginAsync`,
       method: 'post',
       data
     })
@@ -239,7 +239,7 @@ export {
   
   export function GetOrgPageListByAuthCodeAsync(data) {
     return request2({
-      url: `${domain2}Authentication/GetOrgPageListByAuthCodeAsync`,
+      url: `${domain2}/Authentication/GetOrgPageListByAuthCodeAsync`,
       method: 'post',
       data
     })
@@ -247,7 +247,7 @@ export {
   
   export function GetTokenInfoByAuthCode(data) {
     return request2({
-      url: `${domain2}Authentication/GetTokenInfoByAuthCode`,
+      url: `${domain2}/Authentication/GetTokenInfoByAuthCode`,
       method: 'post',
       data
     })
@@ -255,7 +255,7 @@ export {
   
   export function GetSMSCodeAsync(data) {
     return request2({
-      url: `${domain2}Authentication/GetSMSCodeAsync`,
+      url: `${domain2}/Authentication/GetSMSCodeAsync`,
       method: 'post',
       data
     })
@@ -263,14 +263,14 @@ export {
   
   export function LoginWithSMSAsync(data) {
     return request2({
-      url: `${domain2}Authentication/LoginWithSMSAsync`,
+      url: `${domain2}/Authentication/LoginWithSMSAsync`,
       method: 'post',
       data
     })
   }
   export function GetEmpAuthorityListForWebByEmpIdAsync(params) {
     return request2({
-      url: `${domain1}Authorization/GetEmpAuthorityListForWebByEmpIdAsync`,
+      url: `${domain1}/Authorization/GetEmpAuthorityListForWebByEmpIdAsync`,
       method: 'get',
       params
     })
