@@ -117,7 +117,7 @@ export {
 
   export function LoginAsync(data) {
     return request2({
-      url: `${domain2}Authentication/LoginAsync`,
+      url: `${domain2}/Authentication/LoginAsync`,
       method: 'post',
       data
     })
@@ -142,7 +142,7 @@ export {
   
   export function GetOrgPageListByAuthCodeAsync(data) {
     return request2({
-      url: `${domain2}Authentication/GetOrgPageListByAuthCodeAsync`,
+      url: `${domain2}/Authentication/GetOrgPageListByAuthCodeAsync`,
       method: 'post',
       data
     })
@@ -150,7 +150,7 @@ export {
   
   export function GetTokenInfoByAuthCode(data) {
     return request2({
-      url: `${domain2}Authentication/GetTokenInfoByAuthCode`,
+      url: `${domain2}/Authentication/GetTokenInfoByAuthCode`,
       method: 'post',
       data
     })
@@ -158,7 +158,7 @@ export {
   
   export function GetSMSCodeAsync(data) {
     return request2({
-      url: `${domain2}Authentication/GetSMSCodeAsync`,
+      url: `${domain2}/Authentication/GetSMSCodeAsync`,
       method: 'post',
       data
     })
@@ -166,7 +166,7 @@ export {
   
   export function LoginWithSMS(data) {
     return request2({
-      url: `${domain2}Authentication/LoginWithSMSAsync`,
+      url: `${domain2}/Authentication/LoginWithSMSAsync`,
       method: 'post',
       data
     })
@@ -174,7 +174,7 @@ export {
   
   export function GetEmpAuthorityListForWebByEmpIdAsync(params) {
     return request2({
-      url: `${domain1}Authorization/GetEmpAuthorityListForWebByEmpIdAsync`,
+      url: `${domain1}/Authorization/GetEmpAuthorityListForWebByEmpIdAsync`,
       method: 'get',
       params
     })
